feat: migrate IPC and HTTP discovery tests to rstest integration tests
- Migrated ipc_message_queue tests to tests/ipc_message_queue_integration_tests.rs
- Migrated ipc_batch tests to tests/ipc_batch_integration_tests.rs (skipped on Windows)
- Migrated http_discovery tests to tests/http_discovery_integration_tests.rs
- Made ipc module public for integration test access
- Fixed Timer::should_tick() visibility with test-helpers feature
- Updated justfile to include new integration tests
- All 41 Rust integration tests + 40 Python tests passing
- Code coverage: 53.76% (exceeds 50% requirement)
