+++ src/ipc/message_queue.rs
@@ -362,56 +362,8 @@ impl Default for MessageQueue {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn small_queue(block_on_full: bool, max_retries: u32) -> MessageQueue {
-        let cfg = MessageQueueConfig {
-            capacity: 1,
-            block_on_full,
-            max_retries,
-            retry_delay_ms: 1,
-        };
-        MessageQueue::with_config(cfg)
-    }
-
-    #[test]
-    fn test_push_pop_and_metrics() {
-        let q = small_queue(false, 0);
-        q.push(WebViewMessage::EvalJs("1+1".into()));
-        let mut processed = 0;
-        processed += q.process_all(|m| match m {
-            WebViewMessage::EvalJs(s) => assert_eq!(s, "1+1"),
-            _ => unreachable!(),
-        });
-        assert_eq!(processed, 1);
-        let snap = q.get_metrics_snapshot();
-        assert_eq!(snap.messages_sent, 1);
-        assert_eq!(snap.messages_received, 1);
-        assert!(snap.peak_queue_length >= 1);
-    }
-
-    #[test]
-    fn test_backpressure_drop_and_retry_to_dlq() {
-        let q = small_queue(false, 1);
-        // Fill the queue
-        q.push(WebViewMessage::EvalJs("a".into()));
-        // This immediate push should drop due to full queue
-        q.push(WebViewMessage::EvalJs("b".into()));
-        let snap = q.get_metrics_snapshot();
-        assert!(snap.messages_dropped >= 1);
-
-        // push_with_retry should attempt and then go to DLQ
-        let err = q
-            .push_with_retry(WebViewMessage::EvalJs("c".into()))
-            .unwrap_err();
-        assert!(err.contains("queue full") || err.contains("Channel disconnected"));
-        let stats = q.get_dlq_stats();
-        assert!(stats.total >= 1);
-        assert!(stats.queue_full >= 1 || stats.disconnected >= 1);
-
-        // Drain queue to keep later tests stable
-        let _ = q.process_all(|_| {});
-    }
-}
+// Note: Integration tests have been moved to tests/ipc_message_queue_integration_tests.rs
+// This includes tests for:
+// - Push/pop operations and metrics tracking
+// - Backpressure handling and message dropping
+// - Retry logic and dead letter queue integration
