+++ src/service_discovery/http_discovery.rs
@@ -181,89 +181,10 @@ impl Drop for HttpDiscovery {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_http_discovery_start_stop_and_request() {
-        let mut server = HttpDiscovery::new(0, 9101); // Use port 0 for OS-assigned port
-        assert!(!server.is_running());
-        server.start().await.expect("server should start");
-        assert!(server.is_running());
-
-        // Get the actual bound port from the server
-        let bound_port = server.port;
-        assert!(bound_port > 0, "Port should be set after start");
-
-        // Query the discovery endpoint using reqwest
-        let client = reqwest::Client::new();
-        let resp = client
-            .get(format!("http://127.0.0.1:{}/discover", bound_port))
-            .send()
-            .await
-            .expect("GET /discover should succeed");
-        assert!(resp.status().is_success());
-        let json: DiscoveryResponse = resp.json().await.unwrap();
-        assert_eq!(json.service, "AuroraView Bridge");
-        assert_eq!(json.port, 9101);
-        assert_eq!(json.protocol, "websocket");
-
-        // Stop server
-        server.stop().expect("server should stop");
-        assert!(!server.is_running());
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_http_discovery_double_start_and_double_stop_are_safe() {
-        let mut server = HttpDiscovery::new(0, 9102); // 0 → OS-assign port
-                                                      // First start
-        server.start().await.expect("first start ok");
-        assert!(server.is_running());
-        // Second start should be a no-op and Ok
-        server.start().await.expect("second start ok");
-        assert!(server.is_running());
-        // First stop
-        server.stop().expect("first stop ok");
-        assert!(!server.is_running());
-        // Second stop should not panic
-        server.stop().expect("second stop ok");
-        assert!(!server.is_running());
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_http_discovery_unknown_path_returns_404() {
-        // Use port 0 for OS-assigned port to avoid flakiness
-        let mut server = HttpDiscovery::new(0, 9101);
-        server.start().await.expect("server start");
-
-        let bound_port = server.port;
-        assert!(bound_port > 0, "Port should be set after start");
-        let client = reqwest::Client::new();
-        let resp = client
-            .get(format!("http://127.0.0.1:{}/unknown", bound_port))
-            .send()
-            .await
-            .expect("GET should succeed");
-        assert_eq!(resp.status(), reqwest::StatusCode::NOT_FOUND);
-        server.stop().expect("server stop");
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_http_discovery_stop_without_start_is_ok() {
-        let mut server = HttpDiscovery::new(0, 9101);
-        // Calling stop on a non-started server should be a no-op and Ok
-        server.stop().expect("stop ok");
-        assert!(!server.is_running());
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_http_discovery_drop_stops_server() {
-        let mut server = HttpDiscovery::new(0, 9101);
-        server.start().await.expect("start ok");
-        assert!(server.is_running());
-        drop(server); // should invoke Drop::drop -> stop()
-                      // Allow a short moment for graceful shutdown
-        tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-    }
-}
+// Note: Integration tests have been moved to tests/http_discovery_integration_tests.rs
+// This includes tests for:
+// - HTTP server start/stop lifecycle
+// - Discovery endpoint responses
+// - Double start/stop safety
+// - 404 handling for unknown paths
+// - Drop behavior and graceful shutdown
